REF: Move methods that can be shared with new string dtype
